refactor: reorganize gesture system into dedicated src/gesture/ directory
- Move gesture.{h,cpp}, gesture-manager.{h,cpp}, and gesture-action.{h,cpp} to src/gesture/
- Update CMakeLists.txt to reflect new file locations and add gesture include directory
- Fix include paths in cutie-wlc.cpp and glwindow.h for new directory structure
