+++ src/gesture/gesture-manager.cpp
