+++ src/gesture/gesture-manager.h
